requirements.txt
transformers
torchaudio
pyserial
pyaudio
django
torch  
nltk
